+++ CHANGELOG.rst
@@ -4,6 +4,7 @@ CHANGELOG
 UNRELEASED
 ----------
 
+* 🐛 fix: phone number rule will be just valid with international phone numbers formats
 
 1.3.1 (2023-04-30)
 ------------------